refactor(hamcrest): Move all world stuff into same package
