+++ src/main/java/org/mockbukkit/mockbukkit/world/ChunkSnapshotMock.java
@@ -1,4 +1,4 @@
-package org.mockbukkit.mockbukkit.world.chunk;
+package org.mockbukkit.mockbukkit.world;
 
 import com.google.common.base.Preconditions;
 import org.bukkit.ChunkSnapshot;
@@ -7,7 +7,6 @@ import org.bukkit.block.Biome;
 import org.bukkit.block.data.BlockData;
 import org.jetbrains.annotations.NotNull;
 import org.mockbukkit.mockbukkit.exception.UnimplementedOperationException;
-import org.mockbukkit.mockbukkit.world.coordinate.Coordinate;
 
 import java.util.Map;
