+++ src/main/java/org/mockbukkit/mockbukkit/world/Coordinate.java
@@ -1,4 +1,4 @@
-package org.mockbukkit.mockbukkit.world.coordinate;
+package org.mockbukkit.mockbukkit.world;
 
 import org.jetbrains.annotations.NotNull;
